refactor: seperate component and folder, add suspense
