+++ src/components/search-address/style.ts
@@ -1,0 +1,34 @@
+import styled from 'styled-components';
+
+const SearchAddressContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 8px;
+  width: 100%;
+`;
+
+const SearchOpenButton = styled.button`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 8px 16px;
+  width: 100%;
+  height: 40px;
+
+  border: 1px solid green;
+  border-radius: 8px;
+
+  color: ${({ theme }) => theme.colors?.text?.secondary};
+  background: ${({ theme }) => theme.colors?.background?.primary};
+
+  font-size: 16px;
+  line-height: 20px;
+`;
+
+const CurrentAddressText = styled.span`
+  font-size: 16px;
+  line-height: 20px;
+`;
+
+export { SearchAddressContainer, SearchOpenButton, CurrentAddressText };
